convert ints to double precision values for heuristics

+++ src/main/java/app/candycrisis/search/functions/CostFunction.java
@@ -2,6 +2,6 @@ package app.candycrisis.search.functions;
 
 public interface CostFunction<S, A> {
 
-    public int evaluate(S state, A action);
+    public double evaluate(S state, A action);
 
 }
